feat: try to combine deployAccount and fetchAccount
Having a single interface with options prevents these from
diverging further
