+++ packages/sdk/src/client/index.ts
@@ -1,3 +1,1 @@
-export { deployAccount as deployEcdsaAccount, fetchAccount as fetchEcdsaAccount } from "./ecdsa/actions/account.js";
-export { deployAccount, fetchAccount } from "./passkey/actions/account.js";
 export * from "./session/client.js";
